Change benchmarks to use pydantic classes for evaluate results
Adds strongly typed pydantic interfaces as the return values for all benchmarks to resolve #1948

Currently the `evaluate` function is annotated as `... -> Dict`, but in reality most of the benchmarks return a float. In the case of the `IFEval` it returns a dict, so this PR uses inheritance with pydantic models to support returning the base `overall_accuracy` based interface to the client. However, if they choose to specialize their response into `IFEvalResult` then they have the ability to access the `IFEval` specific fields.

I have manually tested both `IFEval` and random other benchmarks and verified that the code functions properly

An example block that can use this inheritance is something like

```
    def evaluate(
        self, model: BaseRESTModel, logger: BenchmarkLogger | None = None, **kwargs
    ) -> BenchmarkResult | None:
        results: DeepEvalBaseBenchmarkResult = self.benchmark.evaluate(model, **kwargs)
        bm_result = BenchmarkResult(
            overall_score=results.overall_accuracy, individual_scores={}, name=self.get_name()
        )
        return bm_result
```
where `BenchmarkResult` is another pydantic model. Essentially, these changes ensures that the result of each benchmark follows a specified format and also supports serializing to dict/json if the user needs that (hence why a Pydantic model is chosen over a raw dataclass).


It is worth noting that **this is a breaking change**, if that is undesirable then I can convert this PR to have new functions called `evaluate_typed` or something similar. Though in my opinion, it is better for the project long term to enforce every benchmark to only use typed models and warn users in release notes of this change (this could be something done later down the line).

This PR also pins datasets to `<4.0.0` as many benchmarks are reliant on data loader scripts which were deprecated in 4.0.0 causing issues.